tile encoder
