+++ libgdal/src/nativeMain/kotlin/tile/GeomType.kt
@@ -1,4 +1,5 @@
 @file:OptIn(ExperimentalForeignApi::class)
+package tile
 
 import kotlinx.cinterop.ExperimentalForeignApi
 import libgdal.OGRwkbGeometryType
