+++ libgdal/src/nativeMain/kotlin/tile/Tile.kt
@@ -1,3 +1,5 @@
+package tile
+
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.protobuf.ProtoNumber
